pyproject.toml
[project]
name = "data-harvesting"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "argparse>=1.4.0",
    "dvc>=3.63.0",
    "dvclive>=3.48.5",
    "gradysim>=0.7.3",
    "gymnasium>=1.2.0",
    "jupyter>=1.1.1",
    "line-profiler>=5.0.0",
    "matplotlib>=3.10.6",
    "mlflow>=3.4.0",
    "numpy>=2.2.6",
    "pettingzoo==1.24.3",
    "pyyaml>=6.0.2",
    "torch>=2.8.0",
    "torchrl>=0.10.0",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
    "vmas>=1.5.0",
]
